Put circuit-builder into a public function
